Make id detector lintFix more intelligent (#288)
The MatchingId checks that ids start with certain prefix. 

## Problem 

The current fix replaces all the wrong ids with the fileName prefix, loosing the original ids. This fix breaks compilation in majority of the times. After this point it becomes impossible to remember which id was which. 

## Solution

Keep the original id and just add the prefix as a solution. 

It is currently checks if there are casing issues and fixes them too. But it does not fix wrong order issues. `mainMenuSomething` will become `menuMainMainMenuSomething`. If you can think of any simple solution to this, I would be glad to do it. The solutions I came up were either complex or has problems with edge cases.

Fixes #284